+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.1.127 (2023-11-1)
+
+- fix: update variable contents in table
+
+## 0.1.126 (2023-11-1)
+
+- fix: support ctrl/cmd + s open saving dialog
+
+## 0.1.125 (2023-11-1)
+
+- fix: use switch represent public field
+- fix: show tip when no any variables
+- fix: reduce the size of radio
+
 ## 0.1.124 (2023-10-31)
 
 - fix: polish prompt editor styles